J'ai fixé GET_PING.py et je suis en train de faire le client

+++ c2w/protocol/unpacking.py
@@ -13,60 +13,92 @@ def decode(datagram):
     fieldsList = []
     messageHeader = struct.unpack('!BHBH', datagram[:6]) #Contains : messageType, seq_number, user_id, message_length
     fieldsList.append(messageHeader)
+    
+    #PUT_LOGIN
     if messageHeader[0] == 0 : #Login request dataType
         UL = struct.unpack('B', datagram[6:7])
         Username = struct.unpack('!' + str(UL[0]) + 's', datagram[7:])
         fieldsList.append(Username[0].decode('utf-8'))
-        
+    
+    #RESPONSE_LOGIN    
     elif messageHeader[0] == 1 : #Login response dataType
         messageBody = struct.unpack('!BBBH', datagram[6:]) #Struct does not have a 3 byte type, so we use a '!BH' type instead
-        lastEventID = messageBody[2]*math.pow(2,16) + messageBody[3] #For the BH type to be converted into a 3 byte value we need to multiply the single byte type (B) by 2^16
+        lastEventID = int(messageBody[2]*math.pow(2,16) + messageBody[3]) #For the BH type to be converted into a 3 byte value we need to multiply the single byte type (B) by 2^16
         fieldsList.append([messageBody[0], messageBody[1], lastEventID])
+    
+    #PUT_LOGOUT
     elif messageHeader[0] == 2 : #Logout request dataType
         fieldsList.append([])
+    
+    #RESPONSE_LOGOUT
     elif messageHeader[0] == 3 : #Logout response dataType
         messageBody = struct.unpack('!B', datagram[6:])
         fieldsList.append([messageBody[0]])
+    
+    #GET_PING
     elif messageHeader[0] == 4 : #Ping request dataType
         messageBody = struct.unpack('!BHB', datagram[6:])
         lastEventID = messageBody[0]*math.pow(2,16) + messageBody[1]
         fieldsList.append([lastEventID, messageBody[2]])
+    
+    #RESPONSE_PING
     elif messageHeader[0] == 5 : #Ping response dataType
         messageBody = struct.unpack('!BH', datagram[6:])
         lastEventID = messageBody[0]*math.pow(2,16) + messageBody[1]
         fieldsList.append([lastEventID])
+    
+    #GET_EVENTS
     elif messageHeader[0] == 6 : #Events data request dataType
         messageBody = struct.unpack('!BHBB', datagram[6:])
         lastEventID = messageBody[0]*math.pow(2,16) + messageBody[1]
         fieldsList.append([lastEventID, messageBody[2], messageBody[3]])
+    
+    #RESPONSE_EVENTS
     elif messageHeader[0] == 7 : #Events data response dataType
         nbrEvents = struct.unpack('!B', datagram[6:7])
         fieldsList.append(unpacking.decodeEvents(nbreEvents, datagram[7:]))
+    
+    #GET_ROOMS
     elif messageHeader[0] == 8 : #Rooms data request
         messageBody = struct.unpack('!BB', datagram[6:])
         fieldsList.append([messageBody[0], messageBody[1]])
+    
+    #RESPONSE_ROOMS
     elif messageHeader[0] == 9 : #Rooms data response
         nbrRooms = struct.unpack('!B', datagram[6:7])
         fieldsList.append(unpacking.decodeRooms(nbreRooms, datagram[7:]))
+    
+    #GET_USERS
     elif messageHeader[0] == 10 : #Users data request
         messageBody = struct.unpack('!BBB', datagram[6:])
         fieldsList.append([messageBody[0], messageBody[1], messageBody[2]])
+    
+    #RESPONSE_USERS
     elif messageHeader[0] == 11 : #Users reponse request
         nbrUsers = struct.unpack('!B', datagram[6:7])
         fieldsList.append(unpacking.decodeUsers(nbreUsers, datagram[7:]))
+    
+    #PUT_SWITCH_ROOM
     elif messageHeader[0] == 12 : #Switch room request
         messageBody = struct.unpack('!B', datagram[6:])
         fieldsList.append([messageBody[0]])
+    
+    #RESPONSE_SWITCH_ROOM
     elif messageHeader[0] == 13 : #Switch room response 
         messageBody = struct.unpack('!B', datagram[6:])
         fieldsList.append([messageBody[0]])
+    
+    #PUT_NEW_MESSAGE
     elif messageHeader[0] == 14 : #New message request
         messageBody = struct.unpack('!BH', datagram[6:10])
         messageBody += struct.unpack('!'+str(messageBody[1])+'s', datagram[10:])
         fieldsList.append([messageBody[0], messageBody[2]])
+    
+    #RESPONSE_NEW_MESSAGE
     elif messageHeader[0] == 15 : #New_message response
         messageBody = struct.unpack('!B', datagram[6:])
         fieldsList.append([messageBody[0]])
+        
     return fieldsList
     
 def decodeUsers(entryNumber, datagram):
